UnrealBuildTool: Fix Action properties that weren't being serialized
#rnx

[CL 24878293 by joe kirchoff in ue5-main branch]

+++ Engine/Source/Programs/UnrealBuildTool/Executors/ParallelExecutor.cs
@@ -218,8 +218,8 @@ namespace UnrealBuildTool
 
 			var ActiveTasks = new Task<ExecuteResults>[ActualNumParallelProcesses];
 			var ActiveActionIndices = new int[ActualNumParallelProcesses];
-			float MaxParallelProcessWeight = (float)ActualNumParallelProcesses;
-			float ActiveTaskWeight = 0.0f;
+			double MaxParallelProcessWeight = ActualNumParallelProcesses;
+			double ActiveTaskWeight = 0.0;
 			int ActiveTaskCount = 0;
 			bool Success = true;
 			List<float> LoggedCpuUtilization = new List<float>();
